test(video-20810): smoke test for profiles and config applet
