+++ test/e2e/specs/NonESM/profileAndConfigAppletMockedColor.spec.ts
@@ -1,0 +1,12 @@
+import { vpTest } from '../../fixtures/vpTest';
+import { getLinkByName } from '../../testData/pageLinksData';
+import { ExampleLinkName } from '../../testData/ExampleLinkNames';
+import { testProfileAndConfigAppletMockedColors } from '../commonSpecs/profilesAndConfigAppletMockedColors';
+
+const link = getLinkByName(ExampleLinkName.Profiles);
+
+vpTest('Profile and config mocked colors are applied to the players',
+  async ({ page, pomPages }) => {
+    await testProfileAndConfigAppletMockedColors(page, pomPages, link);
+  }
+);
