+++ test/e2e/src/consts/appletRoutes.ts
@@ -1,0 +1,2 @@
+export const PROFILE_ROUTE_PATTERN = '**/_applet_/video_service/video_player_profiles/*.json*';
+export const CONFIG_ROUTE_PATTERN = '**/_applet_/video_service/video_player_config/**/*.json*';
